Refactor: extract ping utility
