+++ src/common/mod.rs
@@ -21,3 +21,4 @@ pub(crate) mod s3;
 pub mod websocket;
 
 pub(crate) mod backoff;
+pub(crate) mod ping;
