+++ src/common/ping.rs
@@ -1,0 +1,29 @@
+use std::{
+    task::{Context, Poll},
+    time::Duration,
+};
+
+use tokio::time;
+
+pub struct PingInterval {
+    interval: Option<time::Interval>,
+}
+
+impl PingInterval {
+    pub fn new(period: Option<u64>) -> Self {
+        Self {
+            interval: period.map(|period| time::interval(Duration::from_secs(period))),
+        }
+    }
+
+    pub fn poll_tick(&mut self, cx: &mut Context<'_>) -> Poll<time::Instant> {
+        match self.interval.as_mut() {
+            Some(interval) => interval.poll_tick(cx),
+            None => Poll::Pending,
+        }
+    }
+
+    pub async fn tick(&mut self) -> time::Instant {
+        std::future::poll_fn(|cx| self.poll_tick(cx)).await
+    }
+}
